Add initial refactors
- Separate Proto and Preferences data stores

- Move Context datastore extensions from singular file to top of respective classes

- Remove @Provides in Core module. @Singleton is sufficient

- Fix dependency insertion in gradle

- Remove data class for storing all primitives. All primitives now inserted individually

- Make preferences read function to accept delegate to prevent writing a different read for each preference stored
